Dropbox 2

+++ personal_assistant/storageapp/views/files_views.py
@@ -1,6 +1,3 @@
-"""
-This module contains views for files.
-"""
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect
 
@@ -12,8 +9,8 @@ class FileViews:
     """
     This class contains views for files.
     """
-    services = FileServices
 
+    services = FileServices
 
     @login_required
     def show_user_files(request):
@@ -22,14 +19,14 @@ class FileViews:
         It takes in a request object and returns an HttpResponse object with the rendered template.
 
         :param request: Get the current user, and then pass it to the render_files_list function in fileviews
-        :return: A list of files that the user has access to
+        :return: render files_list.html in FileServices.render_files_list() with incoming data
         :doc-author: Trelent
         """
+
         return FileViews.services.render_files_list(request)
 
     @login_required
     def delete_file_warning(request, file_id):
-
         """
         The delete_file_warning function is called when the user clicks on the &quot;Delete&quot; button in
         the file_detail.html template.  The function takes a request and a file_id as parameters,
@@ -38,9 +35,10 @@ class FileViews:
 
         :param request: Get the request object, which contains information about the current web request
         :param file_id: Get the file object from the database
-        :return: The deleting_warning
+        :return: render the deleting_warning.html with the context
         :doc-author: Trelent
         """
+
         file = File.objects.get(id=file_id)
         return render(request, 'storageapp/deleting_warning.html', context={'file': file})
 
@@ -52,19 +50,22 @@ class FileViews:
         gets that file from the database and calls services.delete_file() to delete it.
         It then returns files_list which is rendered by render_files_list().
 
+
         :param request: Get the request object, which is used to access information about the current http request
         :param file_id: Identify the file to be deleted
-        :return: The files_list
+        :return: render files_list.html in FileServices.render_files_list() with incoming data
         :doc-author: Trelent
         """
-        file = File.objects.get(id=file_id)
-        FileViews.services.delete_file(file=file)
 
+        if request.method == 'POST':
+            file = File.objects.get(id=file_id)
+            message = FileViews.services.delete_file(file=file)
+
+            return FileViews.services.render_files_list(request, message=message)
         return FileViews.services.render_files_list(request)
 
     @login_required
     def upload_file(request):
-
         """
         The upload_file function is responsible for handling the POST request that comes from the upload_file.html page
         when a user submits a file to be uploaded. The function first checks if there is actually a file attached to the
@@ -73,9 +74,10 @@ class FileViews:
         account and returns its new name (which will be used later when we create an instance of File). We also get some other info
 
         :param request: Get the file from the request
-        :return: The files_list
+        :return: render files_list.html in FileServices.render_files_list() with incoming data
         :doc-author: Trelent
         """
+
         if request.method == 'POST':
             if not request.FILES.get('file'):
                 return render(request, 'storageapp/upload_file.html')
@@ -95,7 +97,6 @@ class FileViews:
 
     @login_required
     def download_file(request, file_id):
-
         """
         The download_file function is called when the user clicks on a file in the list of files.
         It takes in a request and an id for the file to be downloaded, then it gets that file from
@@ -104,11 +105,12 @@ class FileViews:
 
         :param request: Get the request object, which is used to check if the method is post
         :param file_id: Identify the file that is to be downloaded
-        :return: A redirect to the url of the file
+        :return: A redirect to the url of the file if the request method is 'POST'.
+                 Otherwise, it renders the file list.
         :doc-author: Trelent
         """
-        if request.method == 'POST':
 
+        if request.method == 'POST':
             file = File.objects.get(id=file_id)
             url = FileViews.services.download_file(file)
 
@@ -116,20 +118,20 @@ class FileViews:
 
         return FileViews.services.render_files_list(request)
 
-
     @login_required
     def search_by_name(request):
-
         """
         The search_by_name function searches for files by name.
-            It takes a request object as an argument and returns the result of the search_by_name function from FileViews.services module.
+        It takes a request object as an argument and returns the result of the search_by_name function from
+        FileViews.services module.
 
         :param request: Get the user input from the search bar
-        :return: A list of files whose name contains the word entered by the user
+        :return: render files_list.html in FileServices.render_files_list() with incoming data
         :doc-author: Trelent
         """
+
         word = request.GET.get('user_input')
-        message = f'Search result by "{word}":'
+        message = f'Search result for "{word}":'
 
         search_result = File.objects.filter(owner=request.user.id).filter(file_name__contains=word)
